Update packages to remove critical and high security alerts. Update README to improve clarity on how to setup the multi RP simulator.
