+++ multi-rp-simulator/.env.example
@@ -1,0 +1,13 @@
+CLIENT1_DESCRIPTION=
+CLIENT1_TENANT_URI=https://cds-gcsignin-dev.verify.ibm.com/oauth2/.well-known/openid-configuration
+CLIENT1_CLIENT_ID=
+CLIENT1_CLIENT_SECRET=
+CLIENT1_REDIRECT_URI=
+CLIENT1_POST_LOGOUT_REDIRECT_URI=
+
+CLIENT2_DESCRIPTION=
+CLIENT2_TENANT_URI=https://cds-gcsignin-dev.verify.ibm.com/oauth2/.well-known/openid-configuration
+CLIENT2_CLIENT_ID=
+CLIENT2_CLIENT_SECRET=
+CLIENT2_REDIRECT_URI=
+CLIENT2_POST_LOGOUT_REDIRECT_URI=
