+++ multi-rp-simulator/package.json
@@ -22,20 +22,20 @@
   },
   "homepage": "https://github.com/sign-in-canada/OIDC-RP-Simulator#readme",
   "dependencies": {
-    "@haibun/cli": "^0.1.2",
-    "@haibun/web-http": "^0.1.1",
-    "@haibun/web-playwright": "^0.2.2",
+    "@haibun/cli": "^3.6.0",
+    "@haibun/web-http": "^3.6.0",
+    "@haibun/web-playwright": "^3.6.0",
     "@types/express": "^4.17.13",
     "@types/http-errors": "^1.8.1",
     "@types/openid-client": "^3.7.0",
     "dotenv": "^17.2.1",
-    "ejs": "^3.1.6",
-    "express": "^4.21.2",
-    "express-session": "^1.18.2",
+    "ejs": "^4.0.1",
+    "express": "^4.22.1",
+    "express-session": "^1.19.0",
     "http-errors": "^2.0.0",
     "openid-client": "^5.1.1",
-    "passport": "^0.5.3",
-    "ts-node": "^10.4.0"
+    "passport": "^0.7.0",
+    "ts-node": "^10.9.2"
   },
   "devDependencies": {
     "@types/dotenv": "^6.1.1"
